# app/docs/openapi.yaml
openapi: 3.0.0
info:
  title: API Gol Ahora
  version: 1.0.0
  description: Documentación de rutas con control de acceso
servers:
  - url: https://golahora-proyecto-is.onrender.com/api
    description: Producción
  - url: http://localhost/api
    description: Servidor local
  - url: http://{ip}/api
    description: Servidor red local
    variables:
      ip:
        default: 192.168.100.66
        description: Dirección IP del servidor
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: Autenticación vía cookie JWT (plataform=web)
    authorization:
      type: apiKey
      in: header
      name: authorization
      description: Autenticación vía token (plataform=windows)
...
paths:
...
  /img/canchas/{nombre}:
    get:
      summary: Obtener imagen de cancha
      description: Devuelve la imagen de una cancha por nombre de archivo. Si no existe, devuelve una imagen por defecto.
      tags:
        - "Acceso: Público"
        - "Módulo: Gestión de Canchas"
        - "Servicio: Arhivo/Recursos"
      parameters:
        - name: plataform
          in: header
          required: true
          schema:
            type: string
            enum: [web, windows]
          description: Plataforma desde la cual se accede.
        
        - name: nombre
          in: path
          required: true
          schema:
            type: string
          description: Nombre del archivo de la imagen
      responses:
        '200':
          description: Imagen encontrada
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Imagen no encontrada (se devuelve imagen por defecto)
...